requirements.txt
discord.py
jishaku
asyncpg
psycopg2